major pre-release. needs a bit more testing, but was working so far with no known bugs.
deepDebug to enable global debug with browser in each observer etc. (I needed it for two bugs to find, very disturbing debugging shiny apps)

date in report

sc_...Input to create variables in .schnappsEnv

loadInput, not used, since I am now working with defaultValues() for the restore point. This was neeeded because updateInput is not working with modules outside of modules.